+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed issue with discoverers not being found in some cases
+
 ## [1.0.3.3] - 2024-09-10
 
 ### Changed